# src/main/java/com/tradebytes/todo/dto/ErrorResponse.java
package com.tradebytes.todo.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;

/**
 * Standard API error response.
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
@Schema(description = "Standard API error response")
public class ErrorResponse {

    @Schema(description = "When the error occurred", example = "2026-01-11T10:30:00")
    private LocalDateTime timestamp;

    @Schema(description = "HTTP status code", example = "400")
    private int status;

    @Schema(description = "HTTP status reason phrase", example = "Bad Request")
    private String error;

    @Schema(description = "Detailed error message", example = "Validation failed")
    private String message;

    @Schema(description = "Request path that caused the error", example = "/api/todos")
    private String path;

    @Schema(description = "List of field-specific validation errors (only present for validation errors)")
    private List<FieldError> fieldErrors;

    @Data
    @NoArgsConstructor
    @AllArgsConstructor
    @Builder
    @Schema(description = "Field-specific validation error")
    public static class FieldError {

        @Schema(description = "Name of the field that failed validation", example = "description")
        private String field;

        @Schema(description = "Validation error message", example = "Description is required")
        private String message;
    }
}